Fixed third and fourth player input
Fixed third and fourth player input. Currently fourth player input is not practical. All characters jump when the spacebar is pressed. Since we all have controllers now and this game is intended for controller use I am not going to be fixing that. If things change I can fix that easily.
- updated Input Manager Keyboard3Horizontal/Keyboard3Vertical to use jkil instead of 4682 for movement
- updated Input Manager Keyboard4Horizontal/Keyboard4Vertical to use vgbn for input.